+++ src/std_gen.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "std")]
+
 use crate::{Scru128Generator, Scru128Id};
 use once_cell::sync::Lazy;
 use std::sync::Mutex;
@@ -7,6 +9,7 @@ static DEFAULT_GENERATOR: Lazy<Mutex<Scru128Generator>> = Lazy::new(Default::def
 /// Generates a new SCRU128 ID object.
 ///
 /// This function is thread-safe; multiple threads can call it concurrently.
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 pub fn scru128() -> Scru128Id {
     DEFAULT_GENERATOR
         .lock()
@@ -26,6 +29,7 @@ pub fn scru128() -> Scru128Id {
 ///
 /// assert!(regex::Regex::new(r"^[0-9A-Z]{25}$").unwrap().is_match(&x));
 /// ```
+#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
 pub fn scru128_string() -> String {
     scru128().into()
 }